Add return, when count elements <2. Change generic RecursiveTask type to Long

+++ src/main/java/org/example/App.java
@@ -28,7 +28,7 @@ public class App {
 
     }
 
-    private static int sumArray(long[] array) {
+    private static long sumArray(long[] array) {
         ForkJoinPool fjp = new ForkJoinPool();
         ArraySumTask task = new ArraySumTask(array, 0, array.length);
         return fjp.invoke(task);
